refactor signal names
