+++ apps/console/components/tail/signals.ts
